backend/requirements.txt
uwsgi==2.0.22  # or whichever latest version is available
Flask==2.3.3
Flask-Cors==4.0.0
flask-talisman==1.1.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
requests==2.31.0